% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scREhurdle-package.R
\docType{package}
\name{scREhurdle-package}
\alias{scREhurdle-package}
\title{The 'scREhurdle' package.}
\description{
R package implementing the methods from "Detection of differentially expressed genes in discrete single-cell RNA sequencing
data using a hurdle model with correlated random effects".
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
