% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scREhurdle.R
\name{scREhurdle}
\alias{scREhurdle}
\title{Hurdle model with random effects for discrete scRNA-seq data}
\usage{
scREhurdle(Y, treatGroup, useCDR = TRUE, typeRE = "ind", subpop = NULL,
  addCovariates = NULL, coefSamps = c("treatment", "all"),
  parSamps = NULL, adjustMethod = "BH", adapt_engaged = FALSE,
  tol = 1e-04, eta = 0.2, output_samples = 1000, stan_seed = 123, ...)
}
\arguments{
\item{Y}{data.frame or matrix where the rows correspond to genes and
columns correspond to cells. Y must contain integers.}

\item{treatGroup}{vector or factor of length ncol(Y) indicating the 
treatment assigments (control or treatment) of the cells.}

\item{useCDR}{whether CDR, the proportion of genes expressed per cell, is 
calculated and used in model matrix (X).}

\item{typeRE}{set type of random effect to use.  Choices are "none", "ind",
or "corr".}

\item{subpop}{vector or factor of length ncol(Y) indicating the 
cluster/subpopulation assignment of cells. Used ONLY in the 
correlated RE model. Cells in different treatment groups should 
be clustered into separate subpopulations.}

\item{addCovariates}{data.frame containing values of additional covariates to
add to model matrix (X). It is recommended that continuous values are scaled
to have mean 0 and standard deviation 0.5.}

\item{coefSamps}{when set to "treatment", only regression coefficients related to 
determining DE genes (i.e., the treatment indicator coefficients) are stored in the 
results, thereby minimizing object size. The "all" option will store all regression 
coefficients. Note: 2 x nrow(Y) x (ncol(X)+1) parameters are stored with "all".}

\item{parSamps}{character vector indicating additional model parameters to 
store. Choices include: "omega", "phi", "lambda1", "lambda2", "sigma2".
Additional choices for correlated RE model include: "gamma_t", 
"omega_star".}

\item{adjustMethod}{p.adjust.method passed to p.adjust() function.}

\item{adapt_engaged}{passed to vb() function in rstan.}

\item{tol}{passed to vb() function in rstan.}

\item{eta}{passed to vb() function in rstan. Eta values as high as 0.4 may work for some datasets. 
Increasing eta should decrease computational time but could cause errors from Stan. Reduce 
eta if model mispecification errors occur.}

\item{output_samples}{passed to vb() function in rstan.}

\item{stan_seed}{seed for random number generation passed to vb() function in rstan.}

\item{...}{additional parameters to be passed to vb() function in rstan.}
}
\value{
scREhurdle.fit object containing:

\itemize{ 
  \item{stanFit}{: an object of stanfit-class.}
  \item{inputData}{: list of data input into the Stan model.}
  \item{deTab}{: data.frame consisting of beta estimates, statistics, and p-values
        associated with identifying DE genes.}
  \item{geneTab}{: data.frame providing estimates for all stored model parameters 
        corresponding to genes (i.e. beta_L, beta_C, zeta_L, zeta_C, phi).}
  \item{cellTab}{: data.frame providing cellular details and stored model parameter 
        estimates corresponding to cells (i.e. treatGroup, CDR, subpop, omega).}
  \item{hyperEst}{: vector with stored hyperparameter estimates of "lambda1", "lambda2", 
        and/or "sigma2" if included in parSamps.}
}
}
\description{
Wrapper function used to implement the methodology from the
"Detection of differentially expressed genes in discrete-single
cell RNA sequencing data using a hurdle model with correlated
random effects" manuscript.
}
\examples{
\dontrun{
data(toyDat)
gene.names <- rownames(toyDat$scData)
treatment <- substring(colnames(toyDat$scData),1,1)

## DE analysis with independent random effects
exIRE <- scREhurdle(Y=toyDat$scData,treatGroup=treatment,useCDR=TRUE,typeRE = "ind",
coefSamps="treatment",eta=0.4,stan_seed=523)

## DE genes
gene.names[which(exIRE$deTab$chisq.padj <= 0.05)]

## DE analysis with correlated random effects
exCRE <- scREhurdle(Y=toyDat$scData,treatGroup=treatment,useCDR=TRUE,typeRE = "corr",
subpop=toyDat$subpop, coefSamps="treatment",eta=0.4,stan_seed=523)

## DE genes
gene.names[exCRE$deTab$chisq.padj <= 0.05]
}

}
