% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scREhurdle.R
\docType{data}
\name{toyDat}
\alias{toyDat}
\title{Toy scRNA-seq dataset and corresponding clustering assignment}
\format{A list containing the following objects:
 \itemize{
  \item{scDat}{: data.frame with 100 rows (genes) and 50 columns (cells)}
  \item{subpop}{: subpopulation clustering assignment for the 50 cells}
}}
\usage{
toyDat
}
\description{
toyDat contains two objects: scDat and subpop.
The data.frame called "scDat" contains gene expression counts. 25 cells belong to the control (C) group and 25 cells belong to treatment (T) group as indicated
in the column names.
The object "subpop" contains the subpopulation clustering assignment for the cells. The control group is clustered into 2 subpopulations and the treatment 
group is clustered into 3 supopulations.
}
\examples{
data(toyDat)
# Obtain dataset
toyDat$scData

# Obtain clustering assignment
toyDat$subpop
}
\keyword{datasets}
